Renamed sensor
