+++ components/ld2415h/binary_sensor/__init__.py
@@ -7,16 +7,18 @@ from esphome.const import (
 )
 from .. import ld2415h_ns, LD2415HComponent, CONF_LD2415H_ID
 
-LD2415HBinarySensor = ld2415h_ns.class_(
-    "LD2415HBinarySensor", binary_sensor.BinarySensor, cg.Component
+ApproachingBinarySensor = ld2415h_ns.class_(
+    "ApproachingBinarySensor", binary_sensor.BinarySensor
 )
 
 CONFIG_SCHEMA = cv.All(
     cv.COMPONENT_SCHEMA.extend(
         {
-            cv.GenerateID(): cv.declare_id(LD2415HBinarySensor),
+            cv.GenerateID(): cv.declare_id(ApproachingBinarySensor),
             cv.GenerateID(CONF_LD2415H_ID): cv.use_id(LD2415HComponent),
-            cv.Optional(CONF_BINARY_SENSOR): binary_sensor.binary_sensor_schema(),
+            cv.Optional(CONF_BINARY_SENSOR): binary_sensor.binary_sensor_schema(
+                ApproachingBinarySensor,
+            ),
         }
     ),
 )
