+++ components/ld2415h/binary_sensor/approaching_binary_sensor.cpp
@@ -1,14 +1,14 @@
-#include "ld2415h_binary_sensor.h"
+#include "approaching_binary_sensor.h"
 #include "esphome/core/helpers.h"
 #include "esphome/core/log.h"
 
 namespace esphome {
 namespace ld2415h {
 
-static const char *const TAG = "LD2415H.binary_sensor";
+static const char *const TAG = "Approaching.binary_sensor";
 
-void LD2415HSensor::dump_config() {
-  ESP_LOGCONFIG(TAG, "LD2415H BinarySensor:");
+void ApproachingBinarySensor::dump_config() {
+  ESP_LOGCONFIG(TAG, "Approaching BinarySensor:");
   LOG_SENSOR("  ", "Approaching", this->approaching_binary_sensor_);
 }
