+++ components/ld2415h/binary_sensor/approaching_binary_sensor.h
